Moved components to their folders, refactored CreateSetForm
