+++ client/src/App.tsx
@@ -2,15 +2,9 @@ import React from "react";
 
 import "./App.css";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import { Home } from "./components/Home";
+import { Home } from "./components/home-layout/Home";
 import { SetDetails } from "./components/SetDetails";
-import { AddSetForm } from "./components/CreateSetForm";
-
-interface Set {
-  name: string;
-  description: string;
-  _id: string;
-}
+import { AddSetForm } from "./components/createNewSet-layout/CreateSetForm";
 
 export function App() {
   return (
