+++ client/src/components/home-layout/Home.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 
-import "../App.css";
+import "../../App.css";
 import { Link } from "react-router-dom";
 import { SearchBar } from "./SearchBar";
 import { HomeCards } from "./HomeCards";
