+++ client/src/components/home-layout/SearchBar.tsx
